src/main/java/guru/springframework/mapper/CategoryCommandToCategory.java
package guru.springframework.mapper;


import guru.springframework.dto.CategoryDTO;
import guru.springframework.model.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * Created by jt on 6/21/17.
 */
@Component
public class CategoryCommandToCategory implements Converter<CategoryDTO, Category>{

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryDTO source) {
        if (source == null) {
            return null;
        }

        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }
}
